refactored help and info into a separate info package
